remove utils and clean up docstrings
