+++ tests/capture/autograph/test_conditionals.py
@@ -28,8 +28,8 @@ jax = pytest.importorskip("jax")
 # must be below jax importorskip
 from jax.core import eval_jaxpr
 
+from pennylane.capture.autograph.ag_primitives import AutoGraphError
 from pennylane.capture.autograph.transformer import TRANSFORMER, run_autograph
-from pennylane.capture.autograph.utils import AutoGraphError
 
 check_cache = TRANSFORMER.has_cache
